ZMLWR-48: ID revolution
change the way master handles ID.

Got rid of the maps and slices of N order, to save memory.
This change removes the peacemaker, the component which solves conflicts among nodes.
Therefore, from now on, every node is supposed to be fully managed by its master and to being connected when empty and/or not in conflict.